requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

datasets==2.10.1
nltk==3.8.1
numpy==1.21.5
pandas==1.4.1
PyYAML==5.4.1
streamlit==1.20.0
torch==1.11.0
torch-xla==1.0
tqdm==4.62.3
transformers==4.17.0
